+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `SandboxConnection` that contains login information from sandboxes.
 - `display_type()` function for detecting the current display type (e.g. "full", "rich", etc.)
 - Bedrock: redact authentication model args from eval logs.
+- OpenAI: warn when `temperature` is used with o1 models (as it is not supported).
 
 ## v0.3.54 (26 December 2024)